+++ x/rewards/types/message_create_denom_reward_schedule.go
@@ -31,8 +31,16 @@ func (msg *MsgCreateDenomRewardSchedule) ValidateBasic() error {
 		return errorsmod.Wrap(ErrInvalidStakingDenom, "denom cannot be empty")
 	}
 
+	if err := sdk.ValidateDenom(msg.Denom); err != nil {
+		return errorsmod.Wrap(ErrInvalidStakingDenom, err.Error())
+	}
+
 	if msg.PrizeDenom == "" {
 		return errorsmod.Wrap(ErrInvalidPrizeDenom, "prize_denom cannot be empty")
+	}
+
+	if err := sdk.ValidateDenom(msg.PrizeDenom); err != nil {
+		return errorsmod.Wrap(ErrInvalidPrizeDenom, err.Error())
 	}
 
 	if msg.DailyAmount.IsNil() || !msg.DailyAmount.IsPositive() {