+++ x/rewards/types/message_update_denom_reward_schedule.go
@@ -29,6 +29,10 @@ func (msg *MsgUpdateDenomRewardSchedule) ValidateBasic() error {
 		return errorsmod.Wrap(ErrInvalidStakingDenom, "denom cannot be empty")
 	}
 
+	if err := sdk.ValidateDenom(msg.Denom); err != nil {
+		return errorsmod.Wrap(ErrInvalidStakingDenom, err.Error())
+	}
+
 	if msg.ScheduleId == "" {
 		return errorsmod.Wrap(ErrInvalidScheduleId, "schedule_id cannot be empty")
 	}